Steps towards attrack test
