+++ packages/reporting/test/index.js
@@ -18,3 +18,4 @@ import './request/unit/utils-test.spec.js';
 import './request/unit/step-context-test.spec.js';
 import './request/unit/qs-whitelist2-test.spec.js';
 import './request/unit/hash-test.spec.js';
+import './request/unit/index.spec.js';
